Safari-611.1.10/LayoutTests/transitions/transition-end-event-create-expected.txt

Test dynamic creation of WebKitTransitionEnd event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createEvent exists

Test creation of WebKitTransitionEvent
PASS typeof(ev) is 'object'
PASS ev.propertyName is ''
PASS ev.elapsedTime is 0.0

Create a MouseEvent and make sure it doesn't have the WebKitTransitionEnd event properties
PASS typeof(ev) is 'object'
PASS ev.propertyName is undefined
PASS ev.elapsedTime is undefined

PASS successfullyParsed is true

TEST COMPLETE
